+++ plugins/modules/oci_database_exadata_infrastructure_actions.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python
-# Copyright (c) 2020, 2022 Oracle and/or its affiliates.
+# Copyright (c) 2020, 2023 Oracle and/or its affiliates.
 # This software is made available to you under the terms of the GPL 3.0 license or the Apache 2.0 license.
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 # Apache License v2.0
@@ -224,6 +224,21 @@ exadata_infrastructure:
             returned: on success
             type: int
             sample: 56
+        additional_compute_count:
+            description:
+                - The requested number of additional compute servers for the Exadata infrastructure.
+            returned: on success
+            type: int
+            sample: 56
+        additional_compute_system_model:
+            description:
+                - Oracle Exadata System Model specification. The system model determines the amount of compute or storage
+                  server resources available for use. For more information, please see L(System and Shape Configuration
+                  Options],https://docs.oracle.com/en/engineered-systems/exadata-cloud-at-customer/ecccm/ecc-system-config-
+                  options.html#GUID-9E090174-5C57-4EB1-9243-B470F9F10D6B)
+            returned: on success
+            type: str
+            sample: X7
         cloud_control_plane_server1:
             description:
                 - The IP address for the first control plane server.
@@ -505,6 +520,8 @@ exadata_infrastructure:
         "additional_storage_count": 56,
         "activated_storage_count": 56,
         "compute_count": 56,
+        "additional_compute_count": 56,
+        "additional_compute_system_model": "X7",
         "cloud_control_plane_server1": "cloud_control_plane_server1_example",
         "cloud_control_plane_server2": "cloud_control_plane_server2_example",
         "netmask": "netmask_example",