+++ plugins/modules/oci_lockbox_access_request_actions.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python
-# Copyright (c) 2020, 2022 Oracle and/or its affiliates.
+# Copyright (c) 2020, 2023 Oracle and/or its affiliates.
 # This software is made available to you under the terms of the GPL 3.0 license or the Apache 2.0 license.
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 # Apache License v2.0
@@ -23,7 +23,7 @@ module: oci_lockbox_access_request_actions
 short_description: Perform actions on an AccessRequest resource in Oracle Cloud Infrastructure
 description:
     - Perform actions on an AccessRequest resource in Oracle Cloud Infrastructure
-    - Handle the AccessRequest
+    - For I(action=handle), handle the AccessRequest
 version_added: "2.9.0"
 author: Oracle (@oracle)
 options:
@@ -273,70 +273,19 @@ class AccessRequestActionsHelperGen(OCIActionsHelperBase):
             access_request_id=self.module.params.get("access_request_id"),
         )
 
-    def approve(self):
-        action_details = oci_common_utils.convert_input_data_to_model_class(
-            self.module.params, HandleAccessRequestDetails
-        )
-        return oci_wait_utils.call_and_wait(
-            call_fn=self.client.handle_access_request,
-            call_fn_args=(),
-            call_fn_kwargs=dict(
-                access_request_id=self.module.params.get("access_request_id"),
-                handle_access_request_details=action_details,
-            ),
-            waiter_type=oci_wait_utils.WORK_REQUEST_WAITER_KEY,
-            operation="{0}_{1}".format(
-                self.module.params.get("action").upper(),
-                oci_common_utils.ACTION_OPERATION_KEY,
-            ),
-            waiter_client=self.get_waiter_client(),
-            resource_helper=self,
-            wait_for_states=oci_common_utils.get_work_request_completed_states(),
-        )
+    def get_action_fn(self, action):
+        action = action.lower()
+        if action in [
+            "approve",
+            "deny",
+            "revoke",
+            "cancel",
+        ]:
+            self.module.params["action"] = action.upper()
+            return getattr(self, "handle", None)
+        return super(AccessRequestActionsHelperGen, self).get_action_fn(action)
 
-    def deny(self):
-        action_details = oci_common_utils.convert_input_data_to_model_class(
-            self.module.params, HandleAccessRequestDetails
-        )
-        return oci_wait_utils.call_and_wait(
-            call_fn=self.client.handle_access_request,
-            call_fn_args=(),
-            call_fn_kwargs=dict(
-                access_request_id=self.module.params.get("access_request_id"),
-                handle_access_request_details=action_details,
-            ),
-            waiter_type=oci_wait_utils.WORK_REQUEST_WAITER_KEY,
-            operation="{0}_{1}".format(
-                self.module.params.get("action").upper(),
-                oci_common_utils.ACTION_OPERATION_KEY,
-            ),
-            waiter_client=self.get_waiter_client(),
-            resource_helper=self,
-            wait_for_states=oci_common_utils.get_work_request_completed_states(),
-        )
-
-    def revoke(self):
-        action_details = oci_common_utils.convert_input_data_to_model_class(
-            self.module.params, HandleAccessRequestDetails
-        )
-        return oci_wait_utils.call_and_wait(
-            call_fn=self.client.handle_access_request,
-            call_fn_args=(),
-            call_fn_kwargs=dict(
-                access_request_id=self.module.params.get("access_request_id"),
-                handle_access_request_details=action_details,
-            ),
-            waiter_type=oci_wait_utils.WORK_REQUEST_WAITER_KEY,
-            operation="{0}_{1}".format(
-                self.module.params.get("action").upper(),
-                oci_common_utils.ACTION_OPERATION_KEY,
-            ),
-            waiter_client=self.get_waiter_client(),
-            resource_helper=self,
-            wait_for_states=oci_common_utils.get_work_request_completed_states(),
-        )
-
-    def cancel(self):
+    def handle(self):
         action_details = oci_common_utils.convert_input_data_to_model_class(
             self.module.params, HandleAccessRequestDetails
         )